pyproject.toml
[project]
name = "core sim 2"
version = "0.1.0"
description = "Cookiecutter template for generic Python 3 project with tooling configuration for testing, linting and IDE support"
authors = [{ name = "Li Lin", email = "guru.lin@gmail.com" }]
readme = "README.md"
requires-python = ">= 3.12"
dependencies = [
    "python-dotenv >= 1.0.0",
    "sqlalchemy[mypy] >= 2.0.23",
    "sqlalchemy-utils >= 0.41.1",
    "alembic >= 1.13.0",
    "psycopg[binary, pool] >= 3.1.19",
    "fastapi>=0.111.0",
    "newrelic>=9.11.0",
    "uvicorn>=0.30.1",
    "ulid-py>=1.1.0",
]

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
    "pytest >= 7.4.0",
    "pytest-dotenv >= 0.5.2",
    "pytest-mock >= 3.14",
    "pytest-cov >= 4.1",
    "pre-commit >= 3.7.1",
    "ruff >= 0.4.8",
    "mypy >= 1.10.0",
]

[tool.rye.scripts]
pre = "pre-commit run --all-files"

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]
env_files = [".env"]

[tool.coverage.run]
source = ["."]
omit = ["tests/*"]

# see ruff configuration options at
# https://docs.astral.sh/ruff/configuration/

[tool.ruff]
line-length = 120
indent-width = 4
exclude = [".git", "__pycache__", "venv", ".venv"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "C",
    "I",
    "A",
] # Enable checks for PEP8, imports, and other standard rules
ignore = ["E203", "E266"]

[tool.mypy]
plugins = ["sqlalchemy.ext.mypy.plugin"]
warn_return_any = false
warn_unused_configs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
mypy_path = "."



[[tool.mypy.overrides]]
module = [
    "dataclasses",
    "sqlalchemy_utils",
]
ignore_missing_imports = true
